split dvc_interface
speed up splash screen
create Widgets and Dialogs in ui module
